Add Javadocs and Rename Application
Changes:
* Application is renamed from "Chatbot" to "Gudetama" to make the images in the GUI
* Javadocs is added to most of the classes and their methods
