--- src/main/java/chatbot/Launcher.java
@@ -1,10 +1,0 @@
-package chatbot;
-
-import javafx.application.Application;
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Duke.class, args);
-    }
-}
-
