+++ src/main/java/gudetama/Launcher.java
@@ -1,0 +1,13 @@
+package gudetama;
+
+import javafx.application.Application;
+
+/**
+ * This class serves as the launcher for the Duke application.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Duke.class, args);
+    }
+}
+
