+++ src/main/java/gudetama/ui/DialogBox.java
@@ -1,0 +1,68 @@
+package gudetama.ui;
+
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.geometry.Pos;
+import javafx.scene.Node;
+import javafx.scene.control.Label;
+import javafx.scene.image.ImageView;
+import javafx.scene.layout.HBox;
+
+/**
+ * Represents a dialog box
+ */
+public class DialogBox extends HBox {
+
+    private Label text;
+    private ImageView displayPicture;
+
+    /**
+     * Constructor for DialogBox
+     * @param l Label for the dialog box
+     * @param iv Image view for the dialog box
+     */
+    public DialogBox(Label l, ImageView iv) {
+        text = l;
+        displayPicture = iv;
+
+        text.setWrapText(true);
+        displayPicture.setFitWidth(100.0);
+        displayPicture.setFitHeight(100.0);
+
+        this.setAlignment(Pos.TOP_RIGHT);
+        this.getChildren().addAll(text, displayPicture);
+    }
+
+
+    /**
+     * Flips the alignment of the dialog box to display messages from the left
+     */
+    private void flip() {
+        this.setAlignment(Pos.TOP_LEFT);
+        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
+        FXCollections.reverse(tmp);
+        this.getChildren().setAll(tmp);
+    }
+
+    /**
+     * Creates a user dialog box with Label and ImageView that is provided
+     * @param l Label for the user's message
+     * @param iv ImageView for the user's display picture
+     * @return A DialogBox representing the user's message
+     */
+    public static DialogBox getUserDialog(Label l, ImageView iv) {
+        return new DialogBox(l, iv);
+    }
+
+    /**
+     * Creates a Gudetama dialog box with Label and ImageView that is provided
+     * @param l Label for Gudetama's message
+     * @param iv ImageView for Gudetama's display picture
+     * @return A DialogBox representing Gudetama's message
+     */
+    public static DialogBox getGudetamaDialog(Label l, ImageView iv) {
+        var db = new DialogBox(l, iv);
+        db.flip();
+        return db;
+    }
+}
